chips: rp2040: move the DMA driver into the shared rp2xxx crate
The PIO gSPI driver the CYW43439 radio runs over needs DMA, so the
RP2350 needs this driver too.

Most of the block is the same on both chips. The channel registers are
identical, and the interrupt registers start at the same +0x400,
because the RP2350's four extra channels take up exactly the space the
RP2040 leaves reserved there. Its two extra interrupt sets follow the
two shared ones and are not declared, since this driver has only ever
serviced IRQ 0 and IRQ 1.

CTRL_TRIG is the exception and it is the dangerous one. The RP2350 adds
INCR_READ_REV at bit 5 and INCR_WRITE_REV at bit 7, so every field
above bit 4 moves up by two places. Nothing about that is visible to
the compiler: register_bitfields! does not check bit positions the way
register_structs! checks offsets, so a field declared at the wrong bit
builds cleanly and then selects the wrong transfer request on hardware.

Two things keep that from being a latent bug.

First, the fields that moved are not declared in the shared crate at
all. They are built from shifts in a DmaLayout trait that each chip
crate fills in. Only five of them are ever used: INCR_WRITE, CHAIN_TO,
TREQ_SEL, BSWAP and BUSY. SNIFF_EN, IRQ_QUIET, RING_SEL and RING_SIZE
also moved, and are simply not declared, because a field that is never
declared cannot be declared wrongly.

Second, chips/rp2040 carries tests over those five. They check each
field against the bit the RP2040 datasheet gives it, check the two
whole CTRL_TRIG words the gSPI driver writes on every transfer, and
check the transfer request table. Each of the five was confirmed to
fail the tests when set to its RP2350 value.

While here, about 250 lines of register declarations that no code reads
are dropped: the timers, the sniffer, MULTI_CHAN_TRIGGER, CHAN_ABORT,
N_CHANNELS, FIFO_LEVELS and all twelve channels' debug registers. They
would otherwise have to be transcribed for the RP2350 as well, which is
where an unverifiable error would sit quietly.

The channel count becomes a const generic, 12 here and 16 on the
RP2350, and the shared driver takes a channel index. The Channel enum
stays in this crate, since how many channels exist is a fact about the
chip.

    board                  text            bss
    raspberry_pi_pico       +0             +0
    pico_explorer_base      +0             +8
    nano_rp2040_connect     +0             +0
    raspberry_pi_pico_w     +0             +0

text is unchanged everywhere. Dma holds two pointers where it held one,
because the interrupt block's base is now given rather than computed,
which is 4 bytes. Three boards absorb it in padding they already had;
on pico_explorer_base it shows as one symbol, start::BUF, growing from
0x860 to 0x864, and the section rounds to 8.

Two caller-visible changes, both the same shape as 459c6f378's:
Dma::new() becomes dma::new(), and channel() takes an index, so
boards/raspberry_pi_pico_w names Channel0 as usize.

Nothing here was run on RP2040 hardware.

+++ boards/raspberry_pi_pico_w/src/main.rs
@@ -129,7 +129,7 @@ pub unsafe fn start() -> (
     let pio_gspi = PioGspiComponent::new(
         &peripherals.pio0,
         pio::SMNumber::SM0,
-        peripherals.dma.channel(dma::Channel::Channel0),
+        peripherals.dma.channel(dma::Channel::Channel0 as usize),
         dma::Irq::Irq0,
         RPGpio::GPIO29,
         RPGpio::GPIO24,
